layout/style/test/test_computed_style_no_pseudo.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=505515
-->
<head>
  <title>Test for Bug 505515</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <style type="text/css">

  #display { color: black; background: white; }
  #display:first-line { color: blue; }

  </style>
</head>
<body onload="run()">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=505515">Mozilla Bug 505515</a>
<p id="display" style="width: 30em">This <span id="sp">is</span> some text in which the first line is in a different color.</p>
<pre id="test">
<script type="application/javascript">

SimpleTest.waitForExplicitFinish();

/** Test for Bug 505515 **/

function run() {
  var p = document.getElementById("display");
  var span = document.getElementById("sp");

  isnot(span.offsetWidth, 0,
        "span should have width (and we flushed layout)");
  is(getComputedStyle(span, "").color, "rgb(0, 0, 0)",
     "span should be black");
  is(getComputedStyle(p, "").color, "rgb(0, 0, 0)",
     "p should be black too");

  SimpleTest.finish();
}

</script>
</pre>
</body>
</html>
